Apply security fixes #6-20 from full audit
#6  GroupStateManager: restrict removeMember to creator or self only
#7  AndroidKeyStoreLocalKeyStore: narrow catch to GeneralSecurityException only
#8  BrokerConfig: switch DEVELOPMENT to ssl://broker.hivemq.com:8883; warn on insecure broker in MqttTransport
#9  JoinProtocol: remove duplicate GroupUpdateAck (canonical copy in sync package)
#10 MqttTransport: evict E2EE shared secret cache when member is removed
#11 ReplicationManager: remove broken dead-code hasMore pagination
#12 (covered by #4) reconnectAttempts already capped at 12
#13 BackupManager: replace version == 1 with version-range check for forward compat
#14 AvatarHttpServer/AvatarDiscovery/AvatarRepository: add random token auth to LAN avatar fetch
#15 build.gradle: revert security-crypto to alpha06 (no stable 1.1.x exists; MasterKey API requires 1.1.x)
#16 build.gradle: bump compileSdk + targetSdk to 35
#17 MqttTransport: age-evict pending messages before count-capping the queue
#18 DataValidator: add optional BIP-39 wordlist check to validateMnemonic
#19 GroupSyncManager: change GroupSyncMessage/GroupUpdateAck/GroupStateRefreshRequest version fields from Int to Long; remove all .toInt() truncation casts
#20 CrashDetectionMonitor: replace @Volatile lastAlertTime with AtomicLong + CAS to eliminate race condition

+++ app/src/main/java/com/example/familysafety/main/SyncIndicator.kt
@@ -106,7 +106,7 @@ fun SyncIndicator(
 }
 
 @Composable
-private fun SyncedChip(version: Int, modifier: Modifier = Modifier) {
+private fun SyncedChip(version: Long, modifier: Modifier = Modifier) {
     var closed by remember(version) { mutableStateOf(false) }
     val shutterProgress = remember(version) { Animatable(0f) }
